requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Django==4.0.2
django-active-link==0.1.8
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-parler==2.3
django-resized==0.3.11
django-rosetta==0.9.8
django-summernote==0.8.20.0
djangorestframework==3.13.1
html2text==2020.1.16
idna==3.3
packaging==21.3
pep517==0.12.0
Pillow==9.0.1
pip-tools==6.5.1
polib==1.1.1
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
